Day-1/requirements.txt
boto3
python-dotenv
requests
aiohttp
matplotlib